[1.27.08] Fix to inheriting voice attributes over <speak> and <voice> tags.
git-svn-id: https://espeak.svn.sourceforge.net/svnroot/espeak/trunk@63 d46cf337-b52f-0410-862d-fd96e6ae7743

+++ src/voice.h
@@ -40,14 +40,14 @@ typedef struct {
 	int formant_factor;      // adjust nominal formant frequencies by this  because of the voice's pitch (256ths)
 
 	// parameters used by Wavegen
-	int freq[N_PEAKS];    // 100% = 256
-	int height[N_PEAKS];  // 100% = 256
-	int width[N_PEAKS];   // 100% = 256
+	short freq[N_PEAKS];    // 100% = 256
+	short height[N_PEAKS];  // 100% = 256
+	short width[N_PEAKS];   // 100% = 256
 
 	// copies without temporary adjustments from embedded commands
-	int freq2[N_PEAKS];    // 100% = 256
-	int height2[N_PEAKS];  // 100% = 256
-	int width2[N_PEAKS];   // 100% = 256
+	short freq2[N_PEAKS];    // 100% = 256
+	short height2[N_PEAKS];  // 100% = 256
+	short width2[N_PEAKS];   // 100% = 256
 
 	int breath[N_PEAKS];  // amount of breath for each formant. breath[0] indicates whether any are set.
 	int breathw[N_PEAKS];  // width of each breath formant
